Updated analytics for Assembly/Area selection
Assembly selection now adds a payload saying how many assemblies are selected in total, and how many of those have "Unity" in the name (we don't list exact payload names because GDPR and developer privacy). Both assembly and area select now report analytics when the window closes, not just when the Apply button is clicked
