+++ Editor/UI/AreaSelectionWindow.cs
@@ -36,7 +36,7 @@ namespace Unity.ProjectAuditor.Editor.UI
 
         private void OnDestroy()
         {
-            m_ProjectAuditorWindow.SetAreaSelection(m_AreaTable.GetTreeViewSelection());
+            ApplySelection();
         }
 
         public static bool IsOpen()
@@ -74,6 +74,16 @@ namespace Unity.ProjectAuditor.Editor.UI
             m_AreaTable = new MultiSelectionTable(m_TreeViewState, multiColumnHeader, names, selection);
         }
 
+        private void ApplySelection()
+        {
+            var analytic = ProjectAuditorAnalytics.BeginAnalytic();
+            m_ProjectAuditorWindow.SetAreaSelection(m_AreaTable.GetTreeViewSelection());
+
+            var payload = new Dictionary<string, string>();
+            payload["areas"] = m_ProjectAuditorWindow.GetSelectedAreasSummary();
+            ProjectAuditorAnalytics.SendUIButtonEventWithKeyValues(ProjectAuditorAnalytics.UIButton.AreaSelectApply, analytic, payload);
+        }
+
         private void OnGUI()
         {
             EditorGUILayout.BeginVertical(GUILayout.ExpandWidth(true));
@@ -85,12 +95,7 @@ namespace Unity.ProjectAuditor.Editor.UI
             if (GUILayout.Button("Clear", GUILayout.Width(50))) m_AreaTable.ClearSelection();
             if (GUILayout.Button("Apply", GUILayout.Width(50)))
             {
-                var analytic = ProjectAuditorAnalytics.BeginAnalytic();
-                m_ProjectAuditorWindow.SetAreaSelection(m_AreaTable.GetTreeViewSelection());
-
-                var payload = new Dictionary<string, string>();
-                payload["areas"] = m_ProjectAuditorWindow.GetSelectedAreasSummary();
-                ProjectAuditorAnalytics.SendUIButtonEventWithKeyValues(ProjectAuditorAnalytics.UIButton.AreaSelectApply, analytic, payload);
+                ApplySelection();
             }
 
             EditorGUILayout.EndHorizontal();
